+++ src/main/java/ru/serebryannikov/voteforlunch/model/AbstractNamedEntity.java
@@ -7,7 +7,7 @@ public class AbstractNamedEntity extends AbstractBaseEntity {
     protected AbstractNamedEntity() {
     }
 
-    protected AbstractNamedEntity(Long id, String name) {
+    protected AbstractNamedEntity(Integer id, String name) {
         super(id);
         this.name = name;
     }
